requirements.txt
anyio==2.2.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.0.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
cryptography==36.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fonttools==4.25.0
idna==3.3
importlib-metadata==4.8.2
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.4.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
kiwisolver==1.3.1
lightgbm==3.3.1
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numexpr==2.8.1
numpy==1.21.2
olefile==0.46
packaging==21.3
pandas==1.3.5
pandocfilters==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
prometheus-client==0.12.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==2.2.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.0
xgboost==1.5.0
zipp==3.6.0
